feat(evalhub): Enhance EvalHub conversion tests and utility functions
- Added a new test to verify that the v1alpha1 conversion omits the `spec.tenancy` field, ensuring it defaults to multi-tenant mode in v1.
- Refactored the `_is_evalhub_crd_available` function into `utils.py` for better organization and reusability, allowing checks for the EvalHub CRD's availability.
- Cleaned up `conftest.py` by removing the old implementation of the CRD availability check.
