+++ tests/ai_safety/evalhub/single_tenancy/utils.py
@@ -2,7 +2,22 @@
 
 from __future__ import annotations
 
+from kubernetes.dynamic import DynamicClient
+from ocp_resources.custom_resource_definition import CustomResourceDefinition
 from ocp_resources.evalhub import EvalHub
+
+
+def _is_evalhub_crd_available(admin_client: DynamicClient) -> bool:
+    """Check if EvalHub CRD is installed on the cluster."""
+    crd_name = "evalhubs.trustyai.opendatahub.io"
+    try:
+        crd = CustomResourceDefinition(
+            client=admin_client,
+            name=crd_name,
+        )
+        return crd.exists
+    except AttributeError, KeyError:
+        return False
 
 
 class SingleTenantEvalHub(EvalHub):
